add pagination

+++ app/Http/Controllers/Admin/Post/DeleteController.php
@@ -11,6 +11,9 @@ class DeleteController extends Controller
 {
     public function __invoke(Post $post)
     {
+        if(auth()->user()->role !== 0){
+            return redirect()->route('admin.post.index');
+        }
         $post->delete();
         return redirect()->route('admin.post.index');
     }
